requirements.txt
azure-functions
pymupdf
pillow
